+++ src/Commands/Utility/Hastebin.ts
@@ -2,13 +2,9 @@ import { Command } from '../../Structures/Command.js';
 import { Message } from 'discord.js';
 import { RegisterCommand } from '../../Structures/Decorator.js';
 import { pasteAliases } from '../../lib/Backend/Pastes.js';
-import { isDM, isText } from '../../lib/types/Discord.js.js';
+import { GuildSettings } from '../../lib/types/Collections.js';
 
 const keys = ['pastebin', ...pasteAliases.keys()];
-// matches prefix, command, and content
-const r = new RegExp(`^(.*?)(${keys.join('|')}) (.*?)$`, 'i');
-// matches prefix and command (no command)
-const s = new RegExp(`^(.*?)(${keys.join('|')})`, 'i');
 
 @RegisterCommand
 export class kCommand extends Command {
@@ -27,9 +23,18 @@ export class kCommand extends Command {
         );
     }
 
-    async init(message: Message, args: string[]) {
-        // "!pastebin" is a valid use, but would return null with the first regex because it looks for a space after the command
-        const [, /*prefix*/, command, content] = message.content.match(r) ?? message.content.match(s);
+    async init(message: Message, args: string[], settings: GuildSettings) {
+        // matches beginning of a line with the set prefix, 
+        // captures the command name (group 1) up until a space
+        // and then matches everything up to the end (including new lines, because of the "s" flag).
+        const parseReg = new RegExp(`^${settings.prefix}(${keys.join('|')}) (.*?)$`, 'si')
+        const content = message.content.match(parseReg);
+        
+        if (!content)
+            return this.Embed.generic(this, 'Invalid use, could not parse arguments. 😕');
+
+        const [, commandName, text] = content;
+        const command = commandName.toLowerCase();
 
         if (command === 'pastebin' || args.length == 0) 
             return this.Embed.success(`
@@ -37,26 +42,8 @@ export class kCommand extends Command {
             ${keys.map(k => `\`\`${k}\`\``).join(', ')}
             `);
 
-        if (message.reference && message.reference.messageID) { // messageID is an optional param, cannot fetch message without it though
-            const channel = message.guild.channels.cache.get(message.reference.channelID);
-            if (isText(channel) || isDM(channel)) {
-                const { content, url } = await channel.messages.fetch(message.reference.messageID);
-                
-                if (content.length === 0)
-                    return this.Embed.fail(`[Message](${url}) has no text content, sorry.`);
-
-                const paste = pasteAliases.get(command.toLowerCase());
-                const pasteLink = await paste(content);
-
-                if (!pasteLink)
-                    return this.Embed.fail('A server error prevented me from uploading the paste. Try a different server!');
-
-                return this.Embed.success(pasteLink);
-            } 
-        }
-
-        const paste = pasteAliases.get(command.toLowerCase());
-        const pasteLink = await paste(content);
+        const paste = pasteAliases.get(command);
+        const pasteLink = await paste(text);
 
         if (!pasteLink)
             return this.Embed.fail('A server error prevented me from uploading the paste. Try a different server!');